backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
firebase-admin==6.2.0
google-cloud-storage==2.12.0
google-cloud-firestore==2.13.1
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.36.4
google-generativeai==0.3.2
google-cloud-logging==3.8.0
google-cloud-secretmanager==2.16.4
pandas==2.1.1
openpyxl==3.1.2
fpdf==1.7.2
reportlab==4.0.7
requests==2.31.0